Periodically reload catalog
Reload the catalog every 15min in the background. If reloading fails,
the old catalog will be kept in memory. The service will not start if
loading the catalog fails initially.

With this change, Helmi will also execute `helm repo update`,
periodically when updating the catalog.

The period can be configured with the `CATALOG_UPDATE_INTERVAL` env
variable.